requirements.txt
flask
opencv-python
opencv-contrib-python
pillow
numpy
flask-cors